Enable Using SDL instead of EDN (#17)
These changes add the ability to specify a
SDL schema, rather than an EDN file.

Additionally, while testing out these changes, it seems that
lacinia (or some app code) isn't properly filtering subscription
results, at least w/ the SDL example.

A bug shall be logged in the graphql.kit repo for verification,
and if the issue is w/ SDL and/or Lacinia, a bug shall be submitted
"upstream" to that repo.